west/requirements.txt
Flask==3.0.3
pandas==1.5.3
numpy==1.23.5
requests==2.32.3
scikit-learn==1.2.2
matplotlib-base==3.9.4
seaborn==0.13.2
joblib==1.2.0
python-dotenv==1.0.1
werkzeug==3.0.6
altair==4.2.2
plotly==5.24.1
streamlit==1.43.2
dash==2.18.2
yellowbrick==1.5
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
torch==1.13.1
torchvision==0.14.1
category_encoders==2.6.4
imbalanced-learn==0.12.4